% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_seasonadj.R
\name{do_seasonadj}
\alias{do_seasonadj}
\title{Helper function for seasonal adjustment}
\usage{
do_seasonadj(x, date)
}
\value{
seasonally adjusted time series as vector
}
\description{
\emph{do_seasonadj()} is used within \emph{gen_transformations}
}
\examples{
\dontrun{
get_matching()
}
}
