% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_q.imp_functions.R
\name{gen_q.imp_agg}
\alias{gen_q.imp_agg}
\title{Aggregate import volume data}
\usage{
gen_q.imp_agg(data)
}
\value{
A dataframe
}
\description{
\emph{get_q.imp_agg()} aggregates data downloaded via \emph{get_q.impfull_destatis}.
}
\examples{
\dontrun{
get_q.imp_agg()
}
}
