% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_helper_functions.R
\name{get_col_num}
\alias{get_col_num}
\title{Create Lineplot in reactable}
\usage{
get_col_num(
  name,
  width = 100,
  reverse = FALSE,
  transf_name = "12m_ra",
  tooltip = NA
)
}
\value{
plot
}
\description{
Create Lineplot in reactable
}
\examples{
\dontrun{
gen_transformations()
}
}
