% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_helper_functions.R
\name{get_col_series}
\alias{get_col_series}
\title{Create Lineplot in reactable}
\usage{
get_col_series(
  y_vars,
  y_labels,
  name = "2021=100",
  width,
  quarterly = F,
  ifo = F
)
}
\value{
plot
}
\description{
Create Lineplot in reactable
}
\examples{
\dontrun{
gen_transformations()
}
}
