% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataframe_functions.R
\name{get_datatable}
\alias{get_datatable}
\title{Create dataframe for reactable table}
\usage{
get_datatable(data, sectors = c(as.character(10:35), "06"), ...)
}
\arguments{
\item{data}{dataframe created by \emph{get_matching}}

\item{sectors}{Sectors to be included. Default: all}
}
\value{
seasonally adjusted time series as vector
}
\description{
\emph{gen_datatable} creates a datatable with the format of the final reactable.
}
\examples{
\dontrun{
gen_dataframe_level()
}
}
