% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataframe_functions.R
\name{get_datatable_level}
\alias{get_datatable_level}
\title{Create dataframe for reactable table}
\usage{
get_datatable_level(
  data,
  lev = 2,
  sts_transf = c("c12m_nsa"),
  sts_series_transf = "level_sa",
  ifo_transf = c("level_nsa"),
  ifo_series_transf = "level_nsa",
  startdate = "2015-01-01",
  enddate = "2024-12-01",
  select_full = FALSE
)
}
\arguments{
\item{data}{dataframe created by \emph{get_matching}}
}
\value{
seasonally adjusted time series as vector
}
\description{
\emph{gen_datatable_level()} is a subfunction of \emph{gen_datatable()}
}
\examples{
\dontrun{
gen_transformations()
}
}
