% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifo_indu.R
\name{get_ifo_indu}
\alias{get_ifo_indu}
\title{Load and format ifo data}
\usage{
get_ifo_indu()
}
\value{
A dataframe
}
\description{
\emph{get_ifo_indu()} loads data on limiting factors for production in the industry originally provided by the ifo Institut.
The data are downloaded from Macrobond using the package MacrobondAPI (which needs specific rights) and formated to be easily combined with further data.
}
\examples{
\dontrun{
get_ifo_indu()
}
}
