% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_helper_functions.R
\name{get_lineplot}
\alias{get_lineplot}
\title{Create Lineplot in reactable}
\usage{
get_lineplot(
  x,
  y_vars = c("production", "importvolume"),
  y_labels = NULL,
  width = 150,
  quarterly = F,
  ifo = F
)
}
\value{
plot
}
\description{
Create Lineplot in reactable
}
\examples{
\dontrun{
gen_transformations()
}
}
