% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matching.R
\name{get_matching}
\alias{get_matching}
\title{Combining dataframes}
\usage{
get_matching(
  matching_base,
  p.indu,
  q.indu,
  p.serv,
  q.serv,
  pimp,
  qimp,
  limiting.indu
)
}
\value{
A dataframe
}
\description{
\emph{get_matching()} combines all datasets.
}
\examples{
\dontrun{
get_matching()
}
}
