% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Create dataframe including data transformations}
\usage{
get_table(
  data,
  width_col_sec = 200,
  width_col_ppi = 55,
  width_col_q.p = 55,
  width_col_ifo = 80,
  width_plot = 200,
  abbr_ppi = "Producer Price Index weight",
  abbr_price = c("ΔProd: Change Rate Producer Prices", "test2"),
  abbr_output = c("test1", "test2"),
  abbr_ifo = c("test1", "test2"),
  expanded = FALSE,
  sorting = list(ppi.weight = "desc"),
  pagination = FALSE,
  filterable = TRUE,
  searchable = FALSE,
  page_size = 7,
  backgroundColor = "#cdcadc",
  ...
)
}
\arguments{
\item{data}{dataframe created by \emph{get_transformation}}

\item{width_col_sec}{width of sector column}

\item{width_col_ppi}{width of ppi column}

\item{width_col_q.p}{width of output and price numerical columns}

\item{width_col_ifo}{width of ifo numerical columns}

\item{abbr_ppi}{Full text header ppi}
}
\value{
seasonally adjusted time series as vector
}
\description{
\emph{gen_transformations} creates a dataframe including multiple transformations. Takes about 20-30 min.
}
\examples{
\dontrun{
gen_dataframe_level()
}
}
